backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
geopandas
folium
streamlit
rapidfuzz
streamlit-folium
plotly
requests
statsmodels
torch
scipy
networkx
shapely
pyjwt
python-multipart
passlib[bcrypt]
matplotlib
tqdm
pymongo
motor
python-dotenv
email-validator
argon2-cffi
reportlab